refactor(struct): extract EPUB package and receipt tests
