+++ src/research/epub.test.ts
@@ -2301,1158 +2301,6 @@ describe('EPUB 3 export', () => {
     expect(content).not.toContain('role="listitem"')
   })
 
-  it('builds a deterministic reflowable container with stored first mimetype', async () => {
-    const first = await buildEpub(paper)
-    const second = await buildEpub(paper)
-    const { files, entries } = inspectEpub(first.bytes)
-
-    expect(first.bytes).toEqual(second.bytes)
-    expect(first.sha256).toBe(second.sha256)
-    expect(first.fileName).toMatch(/\.epub$/)
-    expect(entries[0]).toBe('mimetype')
-    expect(first.bytes[8] | (first.bytes[9] << 8)).toBe(0)
-    expect(strFromU8(files['META-INF/container.xml'])).toContain(
-      'EPUB/package.opf',
-    )
-    const opf = strFromU8(files['EPUB/package.opf'])
-    const navigation = strFromU8(files['EPUB/nav.xhtml'])
-    expect(opf).toContain('version="3.0"')
-    expect(opf).toContain('properties="nav"')
-    expect(opf).toContain('<meta property="schema:accessMode">textual</meta>')
-    expect(opf).toContain(
-      '<meta property="schema:accessibilityFeature">structuralNavigation</meta>',
-    )
-    expect(opf).toContain(
-      '<meta property="schema:accessibilityHazard">none</meta>',
-    )
-    expect(opf).toContain(
-      '<meta property="schema:accessModeSufficient">textual</meta>',
-    )
-    expect(opf).toContain('<meta property="schema:accessibilitySummary">')
-    expect(navigation).toContain('<nav epub:type="toc" role="doc-toc"')
-  })
-
-  it('rejects malformed package graphs and asset bytes that do not match their manifest receipt', async () => {
-    const epub = await buildEpub(paper)
-    const original = unzipSync(epub.bytes)
-    const manifest = JSON.parse(strFromU8(original['EPUB/export.json']))
-    manifest.assets = [
-      {
-        id: 'tampered-asset',
-        href: 'assets/tampered.png',
-        mediaType: 'image/png',
-        kind: 'raster',
-        rendition: 'source-preserved',
-        sha256: '0'.repeat(64),
-      },
-    ]
-    const withAsset = {
-      ...original,
-      'EPUB/export.json': strToU8(`${JSON.stringify(manifest)}\n`),
-      'EPUB/package.opf': strToU8(
-        strFromU8(original['EPUB/package.opf']).replace(
-          '</manifest>',
-          '<item id="tampered-asset" href="assets/tampered.png" media-type="image/png" /></manifest>',
-        ),
-      ),
-      'EPUB/content.xhtml': strToU8(
-        strFromU8(original['EPUB/content.xhtml']).replace(
-          '</main>',
-          '<img src="assets/tampered.png" alt="tampered" /></main>',
-        ),
-      ),
-      'EPUB/assets/tampered.png': new Uint8Array([1, 2, 3]),
-    }
-
-    expect(() => inspectEpub(rezipEpub(withAsset))).toThrow(/SHA-256/i)
-
-    const duplicateOpf = {
-      ...original,
-      'EPUB/package.opf': strToU8(
-        strFromU8(original['EPUB/package.opf']).replace(
-          '</manifest>',
-          '<item id="content" href="content.xhtml" media-type="application/xhtml+xml" /></manifest>',
-        ),
-      ),
-    }
-    expect(() => inspectEpub(rezipEpub(duplicateOpf))).toThrow(
-      /duplicate manifest id/i,
-    )
-
-    const danglingSpine = {
-      ...original,
-      'EPUB/package.opf': strToU8(
-        strFromU8(original['EPUB/package.opf']).replace(
-          'idref="content"',
-          'idref="missing"',
-        ),
-      ),
-    }
-    expect(() => inspectEpub(rezipEpub(danglingSpine))).toThrow(
-      /spine.*missing/i,
-    )
-
-    const malformed = {
-      ...original,
-      'EPUB/content.xhtml': strToU8(
-        strFromU8(original['EPUB/content.xhtml']).replace('</main>', ''),
-      ),
-    }
-    expect(() => inspectEpub(rezipEpub(malformed))).toThrow(/well-formed/i)
-  })
-
-  it('rejects receipt-backed raster signatures without complete image structure', async () => {
-    const epub = await buildEpub(paper)
-    const original = unzipSync(epub.bytes)
-    const cases = [
-      {
-        id: 'truncated-png',
-        href: 'assets/truncated.png',
-        mediaType: 'image/png',
-        bytes: new Uint8Array([0x89, 0x50, 0x4e, 0x47, 0x0d, 0x0a, 0x1a, 0x0a]),
-      },
-      {
-        id: 'truncated-jpeg',
-        href: 'assets/truncated.jpg',
-        mediaType: 'image/jpeg',
-        bytes: new Uint8Array([0xff, 0xd8, 0xff]),
-      },
-      {
-        id: 'truncated-gif',
-        href: 'assets/truncated.gif',
-        mediaType: 'image/gif',
-        bytes: strToU8('GIF89a'),
-      },
-    ] as const
-
-    for (const candidate of cases) {
-      const manifest = JSON.parse(strFromU8(original['EPUB/export.json']))
-      manifest.assets.push({
-        id: candidate.id,
-        href: candidate.href,
-        mediaType: candidate.mediaType,
-        sha256: await sha256Hex(candidate.bytes),
-      })
-      const tampered = {
-        ...original,
-        'EPUB/export.json': strToU8(`${JSON.stringify(manifest)}\n`),
-        'EPUB/package.opf': strToU8(
-          strFromU8(original['EPUB/package.opf']).replace(
-            '</manifest>',
-            `<item id="${candidate.id}" href="${candidate.href}" media-type="${candidate.mediaType}" /></manifest>`,
-          ),
-        ),
-        'EPUB/content.xhtml': strToU8(
-          strFromU8(original['EPUB/content.xhtml']).replace(
-            '</main>',
-            `<img src="${candidate.href}" alt="truncated" /></main>`,
-          ),
-        ),
-        [`EPUB/${candidate.href}`]: candidate.bytes,
-      }
-
-      expect(() => inspectEpub(rezipEpub(tampered))).toThrow(
-        new RegExp(`bytes do not match ${candidate.mediaType}`),
-      )
-    }
-  })
-
-  it('rejects a receipt-backed alternate XHTML document swapped into the OPF spine', async () => {
-    const epub = await buildEpub(paper, getTargetProfile('paperPro'))
-    const files = unzipSync(epub.bytes)
-    const alternate = strToU8(`<?xml version="1.0" encoding="UTF-8"?>
-<html xmlns="http://www.w3.org/1999/xhtml"><head><title>Alternate</title></head><body><main>Alternate spine content.</main></body></html>`)
-    const manifest = JSON.parse(strFromU8(files['EPUB/export.json']))
-    manifest.assets.push({
-      id: 'alternate-content',
-      href: 'alternate.xhtml',
-      mediaType: 'application/xhtml+xml',
-      sha256: await sha256Hex(alternate),
-    })
-    const swapped = {
-      ...files,
-      'EPUB/export.json': strToU8(`${JSON.stringify(manifest)}\n`),
-      'EPUB/package.opf': strToU8(
-        strFromU8(files['EPUB/package.opf'])
-          .replace(
-            '</manifest>',
-            '<item id="alternate-content" href="alternate.xhtml" media-type="application/xhtml+xml" /></manifest>',
-          )
-          .replace('idref="content"', 'idref="alternate-content"'),
-      ),
-      'EPUB/alternate.xhtml': alternate,
-    }
-
-    expect(() =>
-      inspectEpub(rezipEpub(swapped), getTargetProfile('paperPro')),
-    ).toThrow(/spine.*content\.xhtml|content\.xhtml.*spine/i)
-  })
-
-  it('rejects single-quoted dangling href, src, and data attributes', async () => {
-    const epub = await buildEpub(paper)
-    const files = unzipSync(epub.bytes)
-    const cases = [
-      {
-        markup: "<a href='#missing-fragment'>missing</a>",
-        expected: /dangling internal reference #missing-fragment/i,
-      },
-      {
-        markup:
-          "<a href='missing-document.xhtml#missing-fragment'>missing document</a>",
-        expected: /dangling internal reference missing-document\.xhtml/i,
-      },
-      {
-        markup:
-          "<a xmlns:xlink='http://www.w3.org/1999/xlink' xlink:href='#missing-namespaced-fragment'>missing namespaced fragment</a>",
-        expected: /dangling internal reference #missing-namespaced-fragment/i,
-      },
-      {
-        markup: "<img src='assets/missing.png' alt='missing' />",
-        expected: /dangling asset reference assets\/missing\.png/i,
-      },
-      {
-        markup:
-          "<object data='assets/missing.xhtml' type='application/xhtml+xml' />",
-        expected: /dangling asset reference assets\/missing\.xhtml/i,
-      },
-    ]
-
-    for (const candidate of cases) {
-      const tampered = {
-        ...files,
-        'EPUB/content.xhtml': strToU8(
-          strFromU8(files['EPUB/content.xhtml']).replace(
-            '</main>',
-            `${candidate.markup}</main>`,
-          ),
-        ),
-      }
-
-      expect(() => inspectEpub(rezipEpub(tampered))).toThrow(candidate.expected)
-    }
-  })
-
-  it.each([
-    ['supplement.xhtml/', 'supplement.xhtml'],
-    ['chapters//supplement.xhtml', 'chapters/supplement.xhtml'],
-  ])(
-    'does not alias internal href %s to packaged document %s',
-    async (href, packagedHref) => {
-      const epub = await buildEpub(paper)
-      const files = unzipSync(epub.bytes)
-      const packaged = strToU8(`<?xml version="1.0" encoding="UTF-8"?>
-<html xmlns="http://www.w3.org/1999/xhtml"><head><title>Supplement</title></head><body><p>Different exact path.</p></body></html>`)
-      const manifest = JSON.parse(strFromU8(files['EPUB/export.json']))
-      manifest.assets.push({
-        id: 'path-alias-supplement',
-        href: packagedHref,
-        mediaType: 'application/xhtml+xml',
-        sha256: await sha256Hex(packaged),
-      })
-      const tampered = {
-        ...files,
-        'EPUB/export.json': strToU8(`${JSON.stringify(manifest)}\n`),
-        'EPUB/package.opf': strToU8(
-          strFromU8(files['EPUB/package.opf']).replace(
-            '</manifest>',
-            `<item id="path-alias-supplement" href="${packagedHref}" media-type="application/xhtml+xml" /></manifest>`,
-          ),
-        ),
-        'EPUB/content.xhtml': strToU8(
-          strFromU8(files['EPUB/content.xhtml']).replace(
-            '</main>',
-            `<a href="${href}">Broken alias</a></main>`,
-          ),
-        ),
-        [`EPUB/${packagedHref}`]: packaged,
-      }
-
-      expect(() => inspectEpub(rezipEpub(tampered))).toThrow(
-        new RegExp(
-          `dangling internal reference ${href.replaceAll('/', '\\/')}`,
-        ),
-      )
-    },
-  )
-
-  it('rejects hidden duplicate canonical citation text in serialized XHTML', async () => {
-    const citationPaper = structuredClone(paper)
-    citationPaper.nodes = [
-      {
-        id: 'claim',
-        type: 'paragraph',
-        text: 'Prior work [62, 63].',
-        inlineRuns: [
-          {
-            start: 11,
-            end: 19,
-            relationshipId: 'citation-list',
-            semanticRole: 'citation',
-            targetIds: ['reference-62', 'reference-63'],
-          },
-        ],
-        source: 'serialized-citation-audit',
-      },
-      ...[62, 63].map((ordinal) => ({
-        id: `reference-${ordinal}`,
-        type: 'paragraph' as const,
-        text: `Reference ${ordinal}.`,
-        source: 'serialized-citation-audit',
-      })),
-    ]
-    const epub = await buildEpub(citationPaper)
-    const files = unzipSync(epub.bytes)
-    const content = strFromU8(files['EPUB/content.xhtml'])
-    const tamperedContent = content.replace(
-      '</span>.</p>',
-      '<a href="#reference-63" epub:type="biblioref" class="additional-biblioref"><span class="visually-hidden">[62, 63]</span></a></span>.</p>',
-    )
-    expect(tamperedContent).not.toBe(content)
-
-    expect(() =>
-      inspectEpub(
-        rezipEpub({
-          ...files,
-          'EPUB/content.xhtml': strToU8(tamperedContent),
-        }),
-      ),
-    ).toThrow(/duplicate canonical citation text|hidden semantic text/iu)
-  })
-
-  it.each([
-    {
-      label: 'unbounded scholarly destination text',
-      mutate: (content: string) =>
-        content.replace('>Eq. (3)</a>', '>Eq. (3) and the trailing clause</a>'),
-      expected: /unbounded scholarly destination text/iu,
-    },
-    {
-      label: 'overlapping nested destination anchor',
-      mutate: (content: string) =>
-        content.replace(
-          '>Eq. (3)</a>',
-          '><a href="#equation-3">Eq. (3)</a></a>',
-        ),
-      expected: /overlapping.*anchor|nested.*anchor/iu,
-    },
-  ])('rejects $label in serialized XHTML', async ({ mutate, expected }) => {
-    const linkedPaper = structuredClone(paper)
-    linkedPaper.nodes = [
-      {
-        id: 'claim',
-        type: 'paragraph',
-        text: 'See Eq. (3).',
-        inlineRuns: [
-          {
-            start: 4,
-            end: 11,
-            href: '#equation-3',
-            annotationId: 'pdf-link-equation-3',
-          },
-        ],
-        source: 'serialized-link-audit',
-      },
-      {
-        id: 'equation-3',
-        type: 'figure',
-        objectType: 'equation',
-        title: 'Equation 3',
-        relationships: { caption: 'equation-3-caption' },
-        source: 'serialized-link-audit',
-      },
-      {
-        id: 'equation-3-caption',
-        type: 'caption',
-        text: 'Equation 3.',
-        source: 'serialized-link-audit',
-      },
-    ]
-    const epub = await buildEpub(linkedPaper)
-    const files = unzipSync(epub.bytes)
-    const content = strFromU8(files['EPUB/content.xhtml'])
-    const tamperedContent = mutate(content)
-    expect(tamperedContent).not.toBe(content)
-
-    expect(() =>
-      inspectEpub(
-        rezipEpub({
-          ...files,
-          'EPUB/content.xhtml': strToU8(tamperedContent),
-        }),
-      ),
-    ).toThrow(expected)
-  })
-
-  it('preserves canonical reading order and addressable node IDs in XHTML', async () => {
-    const epub = await buildEpub(paper)
-    const { files } = inspectEpub(epub.bytes)
-    const content = strFromU8(files['EPUB/content.xhtml'])
-    const manifest = JSON.parse(strFromU8(files['EPUB/export.json']))
-
-    expect(manifest.canonicalNodeIds).toEqual(
-      paper.nodes.map((node) => node.id),
-    )
-    for (const node of paper.nodes) {
-      expect(content).toContain(`id="${node.id}"`)
-    }
-    expect(content.indexOf(paper.nodes[0].id)).toBeLessThan(
-      content.indexOf(paper.nodes.at(-1)!.id),
-    )
-  })
-
-  it('fails before XHTML generation when canonical node ids are not globally unique', async () => {
-    const duplicateIds = structuredClone(paper)
-    duplicateIds.nodes[1].id = duplicateIds.nodes[0].id
-
-    await expect(buildEpub(duplicateIds)).rejects.toThrow(
-      /canonical node ids must be globally unique/i,
-    )
-  })
-
-  it('requires a typed, non-empty canonicalNodeIds integrity receipt', async () => {
-    const epub = await buildEpub(paper)
-    const files = unzipSync(epub.bytes)
-    const originalManifest = JSON.parse(
-      strFromU8(files['EPUB/export.json']),
-    ) as Record<string, unknown>
-    const invalidReceipts: unknown[] = [undefined, [], ['valid-id', 7]]
-
-    for (const canonicalNodeIds of invalidReceipts) {
-      const manifest = { ...originalManifest }
-      if (canonicalNodeIds === undefined) delete manifest.canonicalNodeIds
-      else manifest.canonicalNodeIds = canonicalNodeIds
-      const tampered = {
-        ...files,
-        'EPUB/export.json': strToU8(`${JSON.stringify(manifest)}\n`),
-      }
-
-      expect(() => inspectEpub(rezipEpub(tampered))).toThrow(
-        /canonicalNodeIds.*non-empty strings/i,
-      )
-    }
-  })
-
-  it('rejects tampered export receipt identity, hash fields, and unknown fields', async () => {
-    const epub = await buildEpub(paper)
-    const files = unzipSync(epub.bytes)
-    const originalManifest = JSON.parse(
-      strFromU8(files['EPUB/export.json']),
-    ) as Record<string, unknown>
-    const expected = {
-      canonicalPaper: paper,
-      sourceCanonicalPaper: paper,
-    }
-    const cases: Array<{
-      name: string
-      mutate: (manifest: Record<string, unknown>) => void
-      expected: RegExp
-    }> = [
-      {
-        name: 'changed receipt URN',
-        mutate: (manifest) => {
-          manifest.identifier =
-            'urn:srt:substituted:000000000000000000000000:publication'
-        },
-        expected: /identifier.*OPF|OPF.*identifier/i,
-      },
-      {
-        name: 'type-confused canonical hash',
-        mutate: (manifest) => {
-          manifest.canonicalContentSha256 = 7
-        },
-        expected: /canonicalContentSha256.*SHA-256/i,
-      },
-      {
-        name: 'malformed source canonical hash',
-        mutate: (manifest) => {
-          manifest.sourceCanonicalContentSha256 = 'A'.repeat(64)
-        },
-        expected: /sourceCanonicalContentSha256.*SHA-256/i,
-      },
-      {
-        name: 'validly shaped but false canonical hash',
-        mutate: (manifest) => {
-          manifest.canonicalContentSha256 = '0'.repeat(64)
-        },
-        expected: /canonicalContentSha256.*canonical input/i,
-      },
-      {
-        name: 'validly shaped but false source canonical hash',
-        mutate: (manifest) => {
-          manifest.sourceCanonicalContentSha256 = '1'.repeat(64)
-        },
-        expected: /sourceCanonicalContentSha256.*canonical input/i,
-      },
-      {
-        name: 'unknown receipt field',
-        mutate: (manifest) => {
-          manifest.untrustedExtension = true
-        },
-        expected: /unknown export manifest field/i,
-      },
-    ]
-
-    for (const candidate of cases) {
-      const manifest = structuredClone(originalManifest)
-      candidate.mutate(manifest)
-      const tampered = {
-        ...files,
-        'EPUB/export.json': strToU8(`${JSON.stringify(manifest)}\n`),
-      }
-
-      expect(
-        () => inspectEpub(rezipEpub(tampered), undefined, expected),
-        candidate.name,
-      ).toThrow(candidate.expected)
-    }
-
-    const opfIdentifier = String(originalManifest.identifier)
-    const mismatchedOpf = {
-      ...files,
-      'EPUB/package.opf': strToU8(
-        strFromU8(files['EPUB/package.opf']).replace(
-          `>${opfIdentifier}</dc:identifier>`,
-          '>urn:srt:substituted:000000000000000000000000:publication</dc:identifier>',
-        ),
-      ),
-    }
-    expect(() =>
-      inspectEpub(rezipEpub(mismatchedOpf), undefined, expected),
-    ).toThrow(/identifier.*OPF|OPF.*identifier/i)
-  })
-
-  it('validates a source PDF hash receipt and binds an expected source hash', async () => {
-    const { reconstruction } = await staleEquationTranscriptFixture()
-    const sourcePdfSha256 = reconstruction.source.sha256
-    const sourcePaper = reconstruction.paper
-    const epub = await buildEpub(sourcePaper, reconstruction)
-    const files = unzipSync(epub.bytes)
-    const originalManifest = JSON.parse(
-      strFromU8(files['EPUB/export.json']),
-    ) as Record<string, unknown>
-    const expected = {
-      canonicalPaper: sourcePaper,
-      sourceCanonicalPaper: sourcePaper,
-      sourcePdfSha256,
-    }
-
-    expect(originalManifest).toMatchObject({
-      canonicalHyphenDeletionCount: 0,
-      canonicalHyphenDeletionContextCounts: {},
-      canonicalHyphenDeletionLedger: [],
-      canonicalHyphenDeletionLedgerSha256:
-        expect.stringMatching(/^[a-f0-9]{64}$/),
-    })
-    expect(() => inspectEpub(epub.bytes, undefined, expected)).not.toThrow()
-
-    for (const invalid of [7, 'A'.repeat(64), 'abc']) {
-      const manifest = {
-        ...originalManifest,
-        sourcePdfSha256: invalid,
-      }
-      const tampered = {
-        ...files,
-        'EPUB/export.json': strToU8(`${JSON.stringify(manifest)}\n`),
-      }
-      expect(() =>
-        inspectEpub(rezipEpub(tampered), undefined, expected),
-      ).toThrow(/sourcePdfSha256.*SHA-256/i)
-    }
-
-    expect(() =>
-      inspectEpub(epub.bytes, undefined, {
-        ...expected,
-        sourcePdfSha256: 'b'.repeat(64),
-      }),
-    ).toThrow(/sourcePdfSha256.*expected source/i)
-  })
-
-  it('persists an exact validated model-consultation receipt in the PDF export manifest', async () => {
-    const withReceipt = await resolvedVisualModelConsultation()
-    const modelConsultations = withReceipt.modelConsultations!
-
-    const epub = await buildReadableEpub(withReceipt.paper, withReceipt)
-    const { files, manifest } = inspectEpub(epub.bytes, undefined, {
-      canonicalPaper: withReceipt.paper,
-      sourceCanonicalPaper: withReceipt.paper,
-      sourcePdfSha256: withReceipt.source.sha256,
-    })
-    const serializedManifest = JSON.parse(
-      strFromU8(files['EPUB/export.json']),
-    ) as Record<string, unknown>
-
-    expect(serializedManifest.modelConsultations).toEqual(modelConsultations)
-    expect(manifest.modelConsultations).toEqual(modelConsultations)
-  })
-
-  it('refuses to publish model-derived state whose receipt was dropped', async () => {
-    // The receipt was validated only when present, so deleting it laundered
-    // the provenance: STRUCT refuses the same document with
-    // MISSING_MODEL_CONSULTATION_RECEIPT, but the research EPUB — a shipping
-    // export path — published it. A guarantee enforceable on one of two exits
-    // is not a guarantee.
-    const laundered = await resolvedVisualModelConsultation()
-    expect(laundered.modelConsultations).toBeDefined()
-    delete (laundered as { modelConsultations?: unknown }).modelConsultations
-
-    await expect(buildReadableEpub(laundered.paper, laundered)).rejects.toThrow(
-      'MISSING_MODEL_CONSULTATION_RECEIPT',
-    )
-  })
-
-  it('rejects a same-source receipt for a differently resolved PDF at the direct EPUB boundary', async () => {
-    const first = await resolvedVisualModelConsultation(0)
-    const differentlyResolved = await resolvedVisualModelConsultation(1)
-    differentlyResolved.modelConsultations = structuredClone(
-      first.modelConsultations,
-    )
-
-    await expect(
-      buildReadableEpub(differentlyResolved.paper, differentlyResolved),
-    ).rejects.toThrow(/model consultation receipt.*semantic state/i)
-  })
-
-  it('rejects a valid alternate-resolution receipt swapped into a packaged EPUB manifest', async () => {
-    const first = await resolvedVisualModelConsultation(0)
-    const alternate = await resolvedVisualModelConsultation(1)
-    const epub = await buildReadableEpub(first.paper, first)
-    const files = unzipSync(epub.bytes)
-    const manifest = JSON.parse(strFromU8(files['EPUB/export.json'])) as Record<
-      string,
-      unknown
-    >
-    manifest.modelConsultations = alternate.modelConsultations
-
-    expect(() =>
-      inspectEpub(
-        rezipEpub({
-          ...files,
-          'EPUB/export.json': strToU8(`${JSON.stringify(manifest)}\n`),
-        }),
-      ),
-    ).toThrow(/model consultation receipt.*binding/i)
-  })
-
-  it('rejects malformed, pending, and unknown model-consultation receipt data', async () => {
-    const { reconstruction } = await staleEquationTranscriptFixture()
-    const validReceipt = emptyModelConsultationReceipt(reconstruction)
-    const withReceipt = withModelConsultations(reconstruction, validReceipt)
-    const epub = await buildEpub(withReceipt.paper, withReceipt)
-    const files = unzipSync(epub.bytes)
-    const originalManifest = JSON.parse(
-      strFromU8(files['EPUB/export.json']),
-    ) as Record<string, unknown>
-
-    const malformed = structuredClone(originalManifest)
-    ;(
-      (malformed.modelConsultations as ModelFallbackReceipt).metrics as Record<
-        string,
-        unknown
-      >
-    ).untrustedNestedField = true
-    expect(() =>
-      inspectEpub(
-        rezipEpub({
-          ...files,
-          'EPUB/export.json': strToU8(`${JSON.stringify(malformed)}\n`),
-        }),
-      ),
-    ).toThrow(/model consultation receipt/i)
-
-    const unknown = structuredClone(originalManifest)
-    unknown.untrustedExtension = true
-    expect(() =>
-      inspectEpub(
-        rezipEpub({
-          ...files,
-          'EPUB/export.json': strToU8(`${JSON.stringify(unknown)}\n`),
-        }),
-      ),
-    ).toThrow(/unknown export manifest field/i)
-
-    const pendingReceipt = await pendingModelConsultationReceipt(reconstruction)
-    const pending = withModelConsultations(reconstruction, pendingReceipt)
-    await expect(buildEpub(pending.paper, pending)).rejects.toThrow(
-      /pending model consultation/i,
-    )
-    const pendingManifest = {
-      ...originalManifest,
-      modelConsultations: pendingReceipt,
-    }
-    expect(() =>
-      inspectEpub(
-        rezipEpub({
-          ...files,
-          'EPUB/export.json': strToU8(`${JSON.stringify(pendingManifest)}\n`),
-        }),
-      ),
-    ).toThrow(/pending model consultation/i)
-  })
-
-  it('binds a persisted model-consultation receipt to its PDF document and source', async () => {
-    const { reconstruction } = await staleEquationTranscriptFixture()
-    const validReceipt = emptyModelConsultationReceipt(reconstruction)
-
-    for (const receipt of [
-      { ...validReceipt, documentId: 'different-document' },
-      { ...validReceipt, sourceSha256: 'b'.repeat(64) },
-    ]) {
-      const mismatched = withModelConsultations(reconstruction, receipt)
-      await expect(buildEpub(mismatched.paper, mismatched)).rejects.toThrow(
-        /model consultation receipt.*(?:document|source)/i,
-      )
-    }
-
-    const withReceipt = withModelConsultations(reconstruction, validReceipt)
-    const epub = await buildEpub(withReceipt.paper, withReceipt)
-    const files = unzipSync(epub.bytes)
-    const originalManifest = JSON.parse(
-      strFromU8(files['EPUB/export.json']),
-    ) as Record<string, unknown>
-    for (const modelConsultations of [
-      { ...validReceipt, documentId: 'different-document' },
-      { ...validReceipt, sourceSha256: 'b'.repeat(64) },
-    ]) {
-      const tampered = { ...originalManifest, modelConsultations }
-      expect(() =>
-        inspectEpub(
-          rezipEpub({
-            ...files,
-            'EPUB/export.json': strToU8(`${JSON.stringify(tampered)}\n`),
-          }),
-        ),
-      ).toThrow(/model consultation receipt.*(?:document|source)/i)
-    }
-  })
-
-  it('rejects a publication export when a PDF reconstruction omits reassessment evidence', async () => {
-    const partial = {
-      source: {
-        format: 'pdf',
-        sha256: 'a'.repeat(64),
-        fileName: 'partial.pdf',
-        byteLength: 1024,
-      },
-      paper,
-      readiness: { ready: true, blockingDiagnosticCodes: [] },
-      noteRelationships: [],
-      visualRelationships: [],
-      assets: [],
-    } as unknown as PdfReconstruction
-
-    await expect(buildEpub(paper, partial)).rejects.toMatchObject({
-      code: 'INCOMPLETE_RECONSTRUCTION',
-    })
-  })
-
-  it('fails closed when publication reassessment omits or corrupts the semantic-flow ledger receipt', async () => {
-    const complete = await readyExternalHyperlinkFixture()
-    const missingDecisions = structuredClone(
-      complete,
-    ) as Partial<PdfReconstruction>
-    delete missingDecisions.sourceSemanticFlowBoundaryDecisions
-    const missingCount = structuredClone(complete) as Partial<PdfReconstruction>
-    delete missingCount.sourceSemanticFlowBoundaryDecisionCount
-    const malformed = structuredClone(complete) as unknown as {
-      sourceSemanticFlowBoundaryDecisions: unknown[]
-      sourceSemanticFlowBoundaryDecisionCount: number
-    }
-    malformed.sourceSemanticFlowBoundaryDecisions = [null]
-    malformed.sourceSemanticFlowBoundaryDecisionCount = 1
-
-    for (const candidate of [missingDecisions, missingCount, malformed]) {
-      await expect(
-        buildEpub(complete.paper, candidate as PdfReconstruction),
-      ).rejects.toMatchObject({ code: 'INCOMPLETE_RECONSTRUCTION' })
-    }
-  })
-
-  it('persists and validates the PDF semantic-flow boundary receipt', async () => {
-    const complete = await readyExternalHyperlinkFixture()
-    const epub = await buildEpub(complete.paper, complete)
-    const files = unzipSync(epub.bytes)
-    const manifest = JSON.parse(strFromU8(files['EPUB/export.json'])) as {
-      sourceSemanticFlowBoundaryCount: number
-      sourceSemanticFlowBoundaryLedger: unknown[]
-      sourceSemanticFlowBoundaryLedgerSha256: string
-    }
-
-    expect(manifest.sourceSemanticFlowBoundaryCount).toBe(
-      complete.sourceSemanticFlowBoundaryDecisionCount,
-    )
-    expect(manifest.sourceSemanticFlowBoundaryLedger).toHaveLength(
-      complete.sourceSemanticFlowBoundaryDecisionCount,
-    )
-    expect(manifest.sourceSemanticFlowBoundaryLedgerSha256).toMatch(
-      /^[a-f0-9]{64}$/u,
-    )
-    expect(() =>
-      inspectEpub(epub.bytes, undefined, {
-        sourceSemanticFlowBoundaryLedgerSha256:
-          manifest.sourceSemanticFlowBoundaryLedgerSha256,
-      }),
-    ).not.toThrow()
-  })
-
-  it('rejects a canonically hashed default-space semantic-flow manifest record', async () => {
-    const complete = await readyExternalHyperlinkFixture()
-    const epub = await buildEpub(complete.paper, complete)
-    const files = unzipSync(epub.bytes)
-    const manifest = JSON.parse(strFromU8(files['EPUB/export.json'])) as Record<
-      string,
-      unknown
-    >
-    const endpoint = {
-      regionId: '1'.repeat(64),
-      lineId: '2'.repeat(64),
-      runIndex: 0,
-      sourceSequenceIndex: 0,
-      sourceRunSha256: '3'.repeat(64),
-      sourceFragmentId: '4'.repeat(64),
-    }
-    const to = {
-      ...endpoint,
-      regionId: '5'.repeat(64),
-      lineId: '6'.repeat(64),
-      sourceSequenceIndex: 1,
-      sourceRunSha256: '7'.repeat(64),
-      sourceFragmentId: '8'.repeat(64),
-    }
-    const evidence = [
-      'exact-source-sequence-adjacency',
-      'explicit-fragment-lineage',
-      'font-baseline-compatible',
-    ]
-    const id = createHash('sha256')
-      .update(
-        JSON.stringify([
-          'pdf-source-semantic-flow-boundary-v1',
-          1,
-          0,
-          'pdf-text',
-          'inline-stacked-fragment',
-          'space',
-          [
-            endpoint.regionId,
-            endpoint.lineId,
-            endpoint.runIndex,
-            endpoint.sourceSequenceIndex,
-            endpoint.sourceRunSha256,
-            endpoint.sourceFragmentId,
-          ],
-          [
-            to.regionId,
-            to.lineId,
-            to.runIndex,
-            to.sourceSequenceIndex,
-            to.sourceRunSha256,
-            to.sourceFragmentId,
-          ],
-          evidence,
-        ]),
-      )
-      .digest('hex')
-    const records = [
-      {
-        id,
-        page: 1,
-        rotation: 0,
-        method: 'pdf-text',
-        topology: 'inline-stacked-fragment',
-        outcome: 'space',
-        from: endpoint,
-        to,
-        evidenceSha256s: evidence.map((value) =>
-          createHash('sha256').update(value).digest('hex'),
-        ),
-      },
-    ]
-    manifest.sourceSemanticFlowBoundaryCount = 1
-    manifest.sourceSemanticFlowBoundaryLedger = records
-    manifest.sourceSemanticFlowBoundaryLedgerSha256 =
-      canonicalJsonSha256ForTest(records)
-    const tampered = {
-      ...files,
-      'EPUB/export.json': strToU8(`${JSON.stringify(manifest)}\n`),
-    }
-
-    expect(() => inspectEpub(rezipEpub(tampered))).toThrow(
-      /semantic-flow boundary ledger is invalid/iu,
-    )
-  })
-
-  it('fails closed at publication export for missing, malformed, or duplicate canonical hyphen deletion proof', async () => {
-    const complete = await readyCanonicalHyphenDeletionFixture()
-    await expect(buildEpub(complete.paper, complete)).resolves.toMatchObject({
-      mode: 'publication',
-    })
-
-    const malformed = structuredClone(complete)
-    malformed.canonicalHyphenBoundaryDecisions[0].proof.pinnedSplit.index += 1
-    const duplicate = structuredClone(complete)
-    duplicate.canonicalHyphenBoundaryDecisions.push(
-      structuredClone(duplicate.canonicalHyphenBoundaryDecisions[0]),
-    )
-    duplicate.canonicalHyphenBoundaryDecisionCount = 2
-    const missing = structuredClone(complete) as Partial<PdfReconstruction>
-    delete missing.canonicalHyphenBoundaryDecisions
-
-    for (const candidate of [
-      missing,
-      {
-        ...structuredClone(complete),
-        canonicalHyphenBoundaryDecisions: [],
-      },
-      malformed,
-      duplicate,
-    ]) {
-      await expect(
-        buildEpub(complete.paper, candidate as PdfReconstruction),
-      ).rejects.toMatchObject({ code: 'INCOMPLETE_RECONSTRUCTION' })
-    }
-    expect(
-      malformed.regions.find(
-        (region) => region.id === 'canonical-hyphen-from-region',
-      )?.text,
-    ).toBe('Repre-')
-  })
-
-  it('persists and validates the normalized PDF canonical hyphen deletion receipt', async () => {
-    const complete = await readyCanonicalHyphenDeletionFixture()
-    const epub = await buildEpub(complete.paper, complete)
-    const files = unzipSync(epub.bytes)
-    const manifest = JSON.parse(strFromU8(files['EPUB/export.json'])) as Record<
-      string,
-      unknown
-    >
-    const ledger = manifest.canonicalHyphenDeletionLedger as Array<
-      Record<string, unknown>
-    >
-    const ledgerSha256 = manifest.canonicalHyphenDeletionLedgerSha256 as string
-
-    expect(manifest).toMatchObject({
-      canonicalHyphenDeletionCount: 1,
-      canonicalHyphenDeletionContextCounts: {
-        'canonical-flow-continuation': 1,
-      },
-      canonicalHyphenDeletionLedgerSha256:
-        expect.stringMatching(/^[a-f0-9]{64}$/),
-    })
-    expect(ledger).toHaveLength(1)
-    expect(ledger[0]).toMatchObject({
-      id: expect.stringMatching(/^[a-f0-9]{64}$/),
-      proof: {
-        tier: 'exact-same-document',
-        sourceBoundaryProven: true,
-        pinnedWordSha256: expect.stringMatching(/^[a-f0-9]{64}$/),
-        pinnedJoinedFormValid: true,
-        pinnedSplit: {
-          leftSha256: expect.stringMatching(/^[a-f0-9]{64}$/),
-          rightSha256: expect.stringMatching(/^[a-f0-9]{64}$/),
-          index: 5,
-        },
-        splitPointValid: true,
-        exactSameDocumentJoinedFormSha256:
-          expect.stringMatching(/^[a-f0-9]{64}$/),
-        sameDocumentJoinedFormValid: true,
-        hardHyphenFormSha256: expect.stringMatching(/^[a-f0-9]{64}$/),
-        hardHyphenCounterproof: null,
-        model: complete.canonicalHyphenBoundaryDecisions[0].proof.model,
-        evidenceSha256s: expect.arrayContaining([
-          expect.stringMatching(/^[a-f0-9]{64}$/),
-        ]),
-      },
-    })
-    expect(JSON.stringify(ledger)).not.toContain('Representation')
-    expect(JSON.stringify(ledger)).not.toContain('language-scope:en-US->en-US')
-    expect(() =>
-      inspectEpub(epub.bytes, undefined, {
-        canonicalHyphenDeletionLedgerSha256: ledgerSha256,
-      }),
-    ).not.toThrow()
-    const legacyManifest = structuredClone(manifest)
-    legacyManifest.schemaVersion = '1.1.0'
-    const legacyRecords =
-      legacyManifest.canonicalHyphenDeletionLedger as Array<{
-        proof: Record<string, unknown>
-      }>
-    for (const record of legacyRecords) delete record.proof.tier
-    legacyManifest.canonicalHyphenDeletionLedgerSha256 =
-      canonicalJsonSha256ForTest(legacyRecords)
-    expect(() =>
-      inspectEpub(
-        rezipEpub({
-          ...files,
-          'EPUB/export.json': strToU8(`${JSON.stringify(legacyManifest)}\n`),
-        }),
-      ),
-    ).not.toThrow()
-    expect(() =>
-      inspectEpub(epub.bytes, undefined, {
-        canonicalHyphenDeletionLedgerSha256: '0'.repeat(64),
-      }),
-    ).toThrow(/canonicalHyphenDeletionLedgerSha256.*expected PDF ledger/i)
-
-    for (const field of [
-      'canonicalHyphenDeletionCount',
-      'canonicalHyphenDeletionContextCounts',
-      'canonicalHyphenDeletionLedger',
-      'canonicalHyphenDeletionLedgerSha256',
-    ]) {
-      const missing = structuredClone(manifest)
-      delete missing[field]
-      expect(() =>
-        inspectEpub(
-          rezipEpub({
-            ...files,
-            'EPUB/export.json': strToU8(`${JSON.stringify(missing)}\n`),
-          }),
-        ),
-      ).toThrow(/missing.*canonical hyphen deletion receipt/i)
-    }
-
-    const countTampered = {
-      ...structuredClone(manifest),
-      canonicalHyphenDeletionCount: 0,
-    }
-    const duplicate = structuredClone(manifest)
-    ;(duplicate.canonicalHyphenDeletionLedger as unknown[]).push(
-      structuredClone(
-        (duplicate.canonicalHyphenDeletionLedger as unknown[])[0],
-      ),
-    )
-    duplicate.canonicalHyphenDeletionCount = 2
-    duplicate.canonicalHyphenDeletionContextCounts = {
-      'canonical-flow-continuation': 2,
-    }
-    const hashTampered = {
-      ...structuredClone(manifest),
-      canonicalHyphenDeletionLedgerSha256: '0'.repeat(64),
-    }
-    const joinedDigestMismatch = structuredClone(manifest)
-    const joinedDigestMismatchRecords =
-      joinedDigestMismatch.canonicalHyphenDeletionLedger as Array<{
-        proof: { exactSameDocumentJoinedFormSha256: string }
-      }>
-    joinedDigestMismatchRecords[0].proof.exactSameDocumentJoinedFormSha256 =
-      'f'.repeat(64)
-    joinedDigestMismatch.canonicalHyphenDeletionLedgerSha256 =
-      canonicalJsonSha256ForTest(joinedDigestMismatchRecords)
-
-    const missingMandatoryEvidence = structuredClone(manifest)
-    const missingMandatoryEvidenceRecords =
-      missingMandatoryEvidence.canonicalHyphenDeletionLedger as Array<{
-        proof: { evidenceSha256s: string[] }
-      }>
-    missingMandatoryEvidenceRecords[0].proof.evidenceSha256s = ['a'.repeat(64)]
-    missingMandatoryEvidence.canonicalHyphenDeletionLedgerSha256 =
-      canonicalJsonSha256ForTest(missingMandatoryEvidenceRecords)
-
-    const forbiddenCounterproof = structuredClone(manifest)
-    const forbiddenCounterproofRecords =
-      forbiddenCounterproof.canonicalHyphenDeletionLedger as Array<{
-        proof: { evidenceSha256s: string[] }
-      }>
-    forbiddenCounterproofRecords[0].proof.evidenceSha256s.push(
-      canonicalHyphenEvidenceSha256ForTest(
-        'hard-hyphen-form-valid:same-document',
-      ),
-    )
-    forbiddenCounterproofRecords[0].proof.evidenceSha256s.sort()
-    forbiddenCounterproof.canonicalHyphenDeletionLedgerSha256 =
-      canonicalJsonSha256ForTest(forbiddenCounterproofRecords)
-
-    for (const tampered of [
-      countTampered,
-      duplicate,
-      hashTampered,
-      joinedDigestMismatch,
-      missingMandatoryEvidence,
-      forbiddenCounterproof,
-    ]) {
-      expect(() =>
-        inspectEpub(
-          rezipEpub({
-            ...files,
-            'EPUB/export.json': strToU8(`${JSON.stringify(tampered)}\n`),
-          }),
-        ),
-      ).toThrow(/canonical hyphen deletion/iu)
-    }
-  })
-
-  it('persists and validates a derived-affix canonical hyphen receipt without joined-surface assertions', async () => {
-    const complete = await readyDerivedAffixHyphenDeletionFixture()
-    const epub = await buildEpub(complete.paper, complete)
-    const files = unzipSync(epub.bytes)
-    const manifest = JSON.parse(strFromU8(files['EPUB/export.json'])) as Record<
-      string,
-      unknown
-    >
-    const ledger = manifest.canonicalHyphenDeletionLedger as Array<{
-      proof: Record<string, unknown>
-    }>
-
-    expect(manifest).toMatchObject({ schemaVersion: '1.2.0' })
-    expect(ledger).toEqual([
-      expect.objectContaining({
-        proof: {
-          tier: 'same-document-derived-affix',
-          sourceBoundaryProven: true,
-          derivedWordSha256: expect.stringMatching(/^[a-f0-9]{64}$/),
-          productivePrefix: {
-            kind: 'prefix',
-            value: 're',
-            affixClass: 'PFX',
-            flag: 'A',
-            crossProduct: true,
-            affixSha256:
-              '70fe5778717d097ce2f3326baaa5c1e4d2206d81a5a81d3ea8e11c4770806dd5',
-          },
-          baseWordSha256: expect.stringMatching(/^[a-f0-9]{64}$/),
-          derivationBindingSha256: expect.stringMatching(/^[a-f0-9]{64}$/),
-          pinnedBaseWordValid: true,
-          pinnedSplit: {
-            leftSha256: expect.stringMatching(/^[a-f0-9]{64}$/),
-            rightSha256: expect.stringMatching(/^[a-f0-9]{64}$/),
-            index: 11,
-          },
-          splitPointValid: true,
-          exactSameDocumentBaseWordSha256:
-            expect.stringMatching(/^[a-f0-9]{64}$/),
-          sameDocumentBaseWordValid: true,
-          hardHyphenFormSha256: expect.stringMatching(/^[a-f0-9]{64}$/),
-          hardHyphenCounterproof: null,
-          model: complete.canonicalHyphenBoundaryDecisions[0].proof.model,
-          evidenceSha256s: expect.arrayContaining([
-            expect.stringMatching(/^[a-f0-9]{64}$/),
-          ]),
-        },
-      }),
-    ])
-    expect(ledger[0].proof).not.toHaveProperty('pinnedJoinedFormValid')
-    expect(ledger[0].proof).not.toHaveProperty('sameDocumentJoinedFormValid')
-    expect(() => inspectEpub(epub.bytes)).not.toThrow()
-
-    const tampered = structuredClone(manifest)
-    const records = tampered.canonicalHyphenDeletionLedger as Array<{
-      proof: {
-        productivePrefix: { flag: string }
-      }
-    }>
-    records[0].proof.productivePrefix.flag = 'Z'
-    tampered.canonicalHyphenDeletionLedgerSha256 =
-      canonicalJsonSha256ForTest(records)
-    expect(() =>
-      inspectEpub(
-        rezipEpub({
-          ...files,
-          'EPUB/export.json': strToU8(`${JSON.stringify(tampered)}\n`),
-        }),
-      ),
-    ).toThrow(/canonical hyphen deletion/iu)
-  })
-
   it('escapes publication metadata rather than emitting invalid XHTML', async () => {
     const escaped = structuredClone(paper)
     escaped.title = 'Evidence & <meaning>'
